clang-tidy: bugprone-virtual-near-miss
A few virtual functions are called by constructors or destructors, which is
dangerous in C++ (as an overridden virtual impl won't be called, only the
one in the current class).  Fix by either marking the function final or
not calling at all (if possible).

+++ glass/src/libnt/native/cpp/NTField2D.cpp
@@ -147,7 +147,7 @@ void NTField2DModel::GroupModel::ObjectModel::SetPoseImpl(double x, double y,
     return;
   }
   auto origArr = value->GetDoubleArray();
-  if (origArr.size() < static_cast<size_t>((m_index + 1) * 3)) {
+  if (static_cast<int>(origArr.size()) < ((m_index + 1) * 3)) {
     return;
   }
